Deploy a web application in AWS/Kubernetes
This repository contains a step-by-step guide to deploying a web application using AWS Elastic Kubernetes Service (EKS). The project was executed as part of an internship program with Brainwave Matrix Solution. Below, I describe the entire process, the challenges encountered, and how the web application was successfully deployed on Kubernetes.